fix Numba type

+++ nucs/solvers/backtrack_solver.py
@@ -129,7 +129,7 @@ class BacktrackSolver(Solver, QueueSolver):
         self.domains_stk = np.empty((stks_max_height, self.problem.domain_nb, 2), dtype=np.int32)
         self.not_entailed_propagators_stk = np.empty((stks_max_height, self.problem.propagator_nb), dtype=np.bool)
         self.dom_update_stk = np.empty((stks_max_height, 2), dtype=np.uint32)
-        self.stks_top = np.ones((1,), dtype=np.uint16)
+        self.stks_top = np.ones((1,), dtype=np.uint32)
         logger.info(f"The stacks of the choice points have a maximal height of {stks_max_height}")
         cp_init(
             self.domains_stk,
